+++ blksprs/__init__.py
@@ -4,6 +4,7 @@ from blksprs.ops.matmul import matmul
 from blksprs.ops.softmax import softmax
 from blksprs.ops.transpose import transpose
 from blksprs.ops.repeat import repeat, repeat_interleave
+from blksprs.misc.partitioning import split, merge
 
 
 class layout:
@@ -15,7 +16,6 @@ class layout:
 class misc:
     from blksprs.misc.broadcast_ops import broadcast_add, broadcast_sub
     from blksprs.misc.exp import exp
-    from blksprs.misc.partitioning import split, merge
     from blksprs.misc.row_wise import row_wise_sum, row_wise_max, row_wise_add, row_wise_sub