+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.8.2] - 2024-10-31
+
+### Added
+
+- Add ``BlksprsTensor`` wrapper class to indicate block-sparse tensors
+
+## [1.8.1] - 2024-10-29
+
+### Fixed
+
+- Fix ``build_sparsity_layout_fast`` not being exposed
+
 ## [1.8] - 2024-10-28
 
 ### Added